Create Ysh
"Ok, this is going to be just a first cut.  You already know that you're to build a shell this term.  Today, I'm merely going to list the features that you need to "add" to ysh (or a similar program you write).

In addition to features currently support by ysh, you're to support as many of the following as you can manage.  Again, I'll be adding more text in the future.  I'm just putting this out there today so that you can use the class period to start designing (i.e. searching for descriptions of how to do all of these things --- you'll find plenty of suggestions on the internet.  But do NOT use code from the internet without my specific approval.  You DO have approval to use the ysh code.)

So, feature to include (currently not supported by ysh) are:

echo
redirection of standard input and output
pipes, e.g.  ls | wc
background jobs
In addition you'll need to include your bash++ translator to your shell, AND, add a Linux utility to indicate the percentage of CPU usage over the last 24 hours.

Finally, I'll want YOU (well as a group) to choose a new Linux utility to support in your shell.

Annnnnnnnnnnd, I'll give "extra" credit for other omissions you find in ysh other than those listed here."